Make clippy happy

+++ src/memory/ppu/chr_memory.rs
@@ -66,7 +66,7 @@ impl ChrMemory {
         };
 
         let bank_count = chr_memory.bank_count();
-        assert_eq!(u32::from(bank_count) * u32::from(chr_memory.bank_size), chr_memory.raw_memory.size() as u32);
+        assert_eq!(u32::from(bank_count) * u32::from(chr_memory.bank_size), chr_memory.raw_memory.size());
         // Power of 2. FIXME: What's the correct behavior when accessing the high banks? Open bus?
         // assert_eq!(bank_count & (bank_count - 1), 0, "Bank count ({bank_count}) must be a power of 2.");
