+++ src/memory/ppu/vram.rs
@@ -9,7 +9,7 @@ const VRAM_SIZE: usize = 2 * KIBIBYTE as usize;
 const CHUNK_SIZE: usize = KIBIBYTE as usize;
 
 // CIRAM
-pub struct Vram(Box<[u8; VRAM_SIZE as usize]>);
+pub struct Vram(Box<[u8; VRAM_SIZE]>);
 
 impl Vram {
     pub fn new() -> Vram {
